+++ build.ps1
@@ -39,6 +39,7 @@ $subpackages = @{
     'com_csmcpforj'              = 'com_csmcpforj.zip'
     'plg_system_csmcpforj'       = 'plg_system_csmcpforj.zip'
     'plg_webservices_csmcpforj'  = 'plg_webservices_csmcpforj.zip'
+    'plg_system_csmcpforj4seo'   = 'plg_system_csmcpforj4seo.zip'
 }
 
 foreach ($srcName in $subpackages.Keys) {